requirements.txt
keras==2.2.2
scikit-learn==0.19.2
numpy==1.14.5
Pillow==3.1.2

